requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
